+++ src/node.rs
@@ -1,3 +1,4 @@
+mod id_provider;
 mod inner;
 mod iters;
 mod node_data;
@@ -7,6 +8,7 @@ mod serializing;
 
 use std::fmt::Debug;
 
+pub use id_provider::NodeIdProver;
 pub use inner::TreeNode;
 pub use iters::{ancestor_nodes, child_nodes, AncestorNodes, ChildNodes};
 pub use node_data::{Element, NodeData};